Move tracing subscriber setup out of App
